backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]==0.24.0
websockets==12.0
# 统一到 langchain-core 0.3.x 生态，避免版本冲突
langchain-core>=0.3.36,<0.4
langchain-openai==0.2.11
langchain-mcp-adapters==0.1.0
starlette>=0.37.2
python-multipart==0.0.6
python-dotenv>=1.0.0
# YAML 配置支持（Agent 图谱）
pyyaml>=6.0.1
# HTTP客户端支持
aiohttp==3.9.1
# SQLite数据库支持
aiosqlite==0.19.0
# SQL查询构建工具(可选)
sqlalchemy==2.0.23
openpyxl>=3.1.2
python-docx>=1.1.0
xlrd>=2.0.1
\# 金融数据与分析
tushare>=1.2.89
pandas>=2.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4